PorewaterDiffusion

Documentation for PorewaterDiffusion.

PorewaterDiffusion.PorewaterProperty

PorewaterDiffusion.Proposal

PorewaterDiffusion.SedimentColumn

PorewaterDiffusion.AND1B

PorewaterDiffusion.AND2A

PorewaterDiffusion.LR04

PorewaterDiffusion.boundaryconditions

PorewaterDiffusion.constants

PorewaterDiffusion.density

PorewaterDiffusion.diffuseadvectcolumn!

PorewaterDiffusion.diffusionadvection

PorewaterDiffusion.dt_climatetimestep

PorewaterDiffusion.porewaterhistory

PorewaterDiffusion.porewaterhistory!

PorewaterDiffusion.velocity

PorewaterDiffusion.PorewaterProperty — Type

PorewaterProperty(n::Int, [, x])

struct to contain sediment column poperties at each node for a prior timestep o and present timestep p.

Constructor function returns an instance of PorewaterProperty with vectors of length n. Optionally provide a value x <: Number to fill vectors with (otherwise values are undefined).

PorewaterDiffusion.Proposal — Type

Proposals(onset, dfrz, dmlt, sea2frz, frz2mlt)

Mutable custom type to hold proposed parameters.

field	description	units

onset	onset of model	ka

dfrz	freezing rate	m/yr

dmlt	melting rate	m/yr

sea2frz	Benthic δ¹⁸O threshold for subglacial freezing	‰

frz2mlt	Benthic δ¹⁸O threshold for subglacial melting	‰

PorewaterDiffusion.SedimentColumn — Type

SedimentColumn(n::Int, [, Cl, O])

struct to contain PorewaterPropertys for the porewater properties of Cl⁻, δ¹⁸O (O), and density (rho).

Constructor function returns an instance of SedimentColumn with PorewaterProperty vectors of length n. Optionally provide values for Cl, O, and rho (otherwise values are undefined).

see also: PorewaterProperty, density

PorewaterDiffusion.AND1B — Method

AND1B()

Generate a seawater NamedTuple with coretop values of ANDRILL-1B.

see also: seawater

PorewaterDiffusion.AND2A — Method

AND2A()

Generate a [`seawater`](@ref) NamedTuple with coretop values of ANDRILL-2A.

see also: seawater

PorewaterDiffusion.LR04 — Method

LR04()

Load a NamedTuple containing the Liesiecki & Raymo 2004 benthic stack (data, publication) interpolated for 1 ka timesteps and going forward in time from 5.32 Ma.

field	description

t	time (ka)

x	benthic δ¹⁸O (‰)

PorewaterDiffusion.boundaryconditions — Method

function boundaryconditions(Cl, d18O, x, sea2freeze,freeze2melt, meltrate, freezerate, Clsw, d18Osw, dz, dt)

Calculates sediment surface boundary condition for δ¹⁸O (d18O) and Cl⁻, based on the thermodynamic state described by the current current benthic δ¹⁸O value x and the threshold values corresponding to subglacial freezing sea2freeze and subglacial melting freeze2melt.

For melting or freezing states, calculates boundary condition from the assumed meltrate, freezerate, timestep dt, length-step dt, and composition of seawater Clsw and d18Osw.

PorewaterDiffusion.constants — Method

julia constants( ; k, dt, dz, depth )`

Returns a NamedTuple of constants and coefficients used in diffusion history calculations. The input constants and their default values are listed in the table below. From these it calculates a few convenience variables: the number of nodes nz, a range of interiornodes, and the product dtdz.

Using the constants, it calculates the number of nodes as well as temperature-dependent diffusion coefficients used in diffusionadvection, using the temperature-depth paramterization of Morin+ 2010. These are returned as Vectors of length nz, where each cell corresponds to a depth node. The two coefficients are k1 and k2, both of which are follwed with cl or w to denote Cl⁻ or water, respectively: k1cl, k2cl, k1w, and k2w.

field	description	default

k	hydraulic conductivity of sediment column	0.1 m / yr

dt	timestep	10 yrs

dz	node spacing	5 m

depth	sediment column depth	2000 m

PorewaterDiffusion.density — Method

density(chlorinity)

Calculates the density of a water parcel with chlorinity in units g/m³ (rather than kg/m³ for convenience with velocity)

see also: velocity

PorewaterDiffusion.diffuseadvectcolumn! — Method

diffuseadvectcolumn!(sc,k)

Calculate diffusive and advective transport of chlorinity and isotope-traced water through a sediment column described by properties in k, a NamedTuple generated by the constants function.

Overwrites the o and p PorewaterProperty fields of sc – a SedimentColumn with pre-existing conditions of Cl, O, and rho in its o fields.

relies on: velocity, density, diffusionadvection,

see also: SedimentColumn, constants

PorewaterDiffusion.diffusionadvection — Method

diffusionadvection(x,above,below,k1,k2,v,dt,dz)

Calculate the property of a node in a vertical profile given the combined effects of diffusion and advection. Returns the property given initial values for the node x, the overlying node above, the underlying node below, diffusion coefficients k1 and k2, vertical advection velocity v, timestep dt, and lengthstep dz. Alternatively provide the product of v * dt * dz for a minor speed-up.

PorewaterDiffusion.dt_climatetimestep — Method

dt_climatetimestep(katime,dt)

A helper function to calculate the number of diffusion model timesteps dt (in years) in each timestep of the climate timescale katime (in kiloannum). Returns an integer.

PorewaterDiffusion.porewaterhistory! — Method

porewaterhistory!(sc::SedimentColumn, p::Proposal, k::NamedTuple, climhist::NamedTuple, sw::NamedTuple, ka_dt::Int)

In-place version of porewaterhistory, which takes every input as an arg (rather than some defaults as kwargs). It also requires you to provide ka_dt – the number of diffusion timesteps in each thousand-year climate timestep.

see also: porewaterhistory

PorewaterDiffusion.porewaterhistory — Method

porewaterhistory(proposals [; k=constants(), climatehistory=LR04(), seawater=AND2A()])

Calculate the porewater advection-diffusion history of chlorinity and O-isotope-traced water in a sediment column described by properties in k (generated with constants) over a given climatehistory (LR04 by default) and coretop seawater compositions.

The provided proposals (as custom type Proposal) describe the sensitivity and response of the system to climate fluctuations as recorded in climatehistory.

See diffuseadvectcolumn! for the underlying diffusion-advection transport calculations.

see also: porewaterhistory!, Proposal, constants, LR04, seawater

PorewaterDiffusion.velocity — Method

velocity(x, above, k)

Calculate the velocity (m/yr) at a node with density x, given the density of the node above, and the hydraulic conductivity k (m/yr).

see also: density