PorewaterDiffusion

Documentation for PorewaterDiffusion.

PorewaterDiffusion.PorewaterProperty

PorewaterDiffusion.SedimentColumn

PorewaterDiffusion.AND1B

PorewaterDiffusion.AND2A

PorewaterDiffusion.boundaryconditions

PorewaterDiffusion.density

PorewaterDiffusion.diffusionadvection

PorewaterDiffusion.velocity

PorewaterDiffusion.PorewaterProperty — Type

PorewaterProperty(n::Int, [, x])

struct to contain sediment column poperties at each node for a prior timestep o and current timestep i.

Constructor function returns an instance of PorewaterProperty with vectors of length n. Optionally provide a value x <: Number to fill vectors with (otherwise values are undefined).

PorewaterDiffusion.SedimentColumn — Type

SedimentColumn(n::Int, [, Cl, O, rho])

struct to contain PorewaterPropertys for the porewater properties of Cl⁻, δ¹⁸O (O), and density (rho).

Constructor function returns an instance of SedimentColumn with PorewaterProperty vectors of length n. Optionally provide values for Cl, O, and rho (otherwise values are undefined).

see also: PorewaterProperty

PorewaterDiffusion.AND1B — Method

AND1B()

Generate a Seawater instance with coretop values of ANDRILL-1B.

see also: Seawater

PorewaterDiffusion.AND2A — Method

AND2A()

Generate a Seawater instance with coretop values of ANDRILL-2A.

see also: Seawater

PorewaterDiffusion.boundaryconditions — Method

function boundaryconditions(Cl, d18O, x, ocean2freeze,freeze2melt, meltrate, freezerate, seawater, dz, dt)

Calculates sediment surface boundary condition for δ¹⁸O (d18O) and Cl⁻, based on the thermodynamic state described by the current current benthic δ¹⁸O value x and the threshold values corresponding to subglacial freezing ocean2freeze and subglacial melting freeze2melt.

For melting or freezing states, calculates boundary condition from the assumed meltrate, freezerate, timestep dt, length-step dt, and composition of seawater Clsw and d18Osw.

PorewaterDiffusion.density — Method

density(chlorinity)

Calculates the density of a water parcel with chlorinity in units g/m³ (rather than g/m³ for convenience with velocity)

see also: velocity

PorewaterDiffusion.diffusionadvection — Method

diffusionadvection(x,above,below,k1,k2,v,dt,dz)

Calculate the property of a node in a vertical profile given the combined effects of diffusion and advection. Returns the property given initial values for the node x, the overlying node above, the underlying node below, diffusion coefficients k1 and k2, vertical advection velocity v, timestep dt, and lengthstep dz. Alternatively provide the product of v * dt * dz for a minor speed-up.

PorewaterDiffusion.velocity — Method

velocity(x, above, k)

Calculate the velocity (m/yr) at a node with density x, given the density of the node above, and the hydraulic conductivity k (m/yr).

see also: density