PorewaterDiffusion

Documentation for PorewaterDiffusion.

PorewaterDiffusion.ClimateHistory

PorewaterDiffusion.Constants

PorewaterDiffusion.CoreData

PorewaterDiffusion.PorewaterProperty

PorewaterDiffusion.Proposal

PorewaterDiffusion.Seawater

PorewaterDiffusion.SedimentColumn

PorewaterDiffusion.LR04

PorewaterDiffusion.andrill1b

PorewaterDiffusion.andrill2a

PorewaterDiffusion.boundaryconditions

PorewaterDiffusion.density

PorewaterDiffusion.diffuseadvectcolumn!

PorewaterDiffusion.diffusionadvection

PorewaterDiffusion.dt_climatetimestep

PorewaterDiffusion.getproposal

PorewaterDiffusion.linterp

PorewaterDiffusion.loglikelihood

PorewaterDiffusion.mcmurdoshelf

PorewaterDiffusion.mcmurdosound

PorewaterDiffusion.normll

PorewaterDiffusion.normpdf

PorewaterDiffusion.porewaterhistory

PorewaterDiffusion.porewaterhistory!

PorewaterDiffusion.porewatermetropolis

PorewaterDiffusion.stopwatch

PorewaterDiffusion.strictpriors

PorewaterDiffusion.velocity

PorewaterDiffusion.ClimateHistory — Type

ClimateHistory

DataType declared as a shorthand for NamedTuple{(:t, :x, :n), Tuple{StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}, Vector{Float64}, Int64}}

The type generated by LR04, which is used for (perhaps unnecessary) type stability and code readability

see also: LR04

PorewaterDiffusion.Constants — Type

julia Constants( ; k, dt, dz, depth )`

Returns a Constants struct containing constants and coefficients used in diffusion history calculations. The inputs and their default values are listed in the table below. From these it calculates a few convenient variables–-the product dt * dz (dtdz), a Range of node depths z (in m), the number of nodes nz, and the penultimate_node–-as well as temperature-dependent diffusion coefficients used in diffusionadvection, using the temperature-depth paramterization of Morin+ 2010. These are returned as Vectors of length nz, where each cell corresponds to a depth node in z. The two coefficients are k1 and k2, both of which are follwed with cl or w to denote Cl⁻ or water, respectively: k1cl, k2cl, k1w, and k2w.

field	description	default

k	hydraulic conductivity of sediment column	0.1 m / yr

dt	timestep	10 yrs

dz	node spacing	5 m

depth	sediment column depth	2000 m

PorewaterDiffusion.CoreData — Type

CoreData(z, mCl, σCl, mO, σO)

Returns an CoreData struct with sediment core data formatted for porewatermetropolis — (; z, Cl = (mu, sig), O = (mu, sig)). Inputs are Vectors for values of sample depths z (meters below sea floor), measured (mean) chlorinity mCl and 1σ uncertainty σCl, and measured (mean) δ¹⁸O mO and 1σ uncertainties σO.

Vectors must all be of the same length. If Cl and δ¹⁸O sampling is not 1:1, use NaN or anything that is not a subtype of Number (e.g. missing, nothing). While NaN is used internally, CoreData does this conversion for you.

CoreData(z, m, σ, measurement)

Same as above, but for a core with only chlorinity or δ¹⁸O data (the other Vectors are empty to return null values in log-likelihood calculations). Provide sample depths z, measured means m, 1σ uncertainties σ, and the measurement as a symbol, e.g. :Cl or :O.

PorewaterDiffusion.jl comes loaded with convenient functions to generate data for andrill2a from Tracy+ 2010 (doi:10.1130/G30849.1) and andrill1b from Pompilio+ 2007 (https://digitalcommons.unl.edu/andrillrespub/37).

PorewaterDiffusion.PorewaterProperty — Type

PorewaterProperty(n::Int, [, x])

struct to contain sediment column poperties at each node for a prior timestep o and present timestep p.

Constructor function returns an instance of PorewaterProperty with vectors of length n. Optionally provide a value x <: Number to fill vectors with (otherwise values are undefined).

PorewaterDiffusion.Proposal — Type

Proposals(onset, dfrz, dmlt, sea2frz, frz2mlt)

Immutable struct to hold proposed parameters.

field	description	units

onset	onset of model	ka

dfrz	freezing rate	m/yr

dmlt	melting rate	m/yr

sea2frz	Benthic δ¹⁸O threshold for subglacial freezing	‰

frz2mlt	Benthic δ¹⁸O threshold for subglacial melting	‰

PorewaterDiffusion.Seawater — Type

Seawater

DataType declared as a shorthand for NamedTuple{(:Cl, :O), Tuple{Float64,Float64}} because Graham likes the splatting functionality of NamedTuples and structs don't have it!

see also: seawater

PorewaterDiffusion.SedimentColumn — Type

SedimentColumn(n::Int, [, Cl, O])

struct to contain PorewaterPropertys for the porewater properties of Cl⁻, δ¹⁸O (O), and density (rho).

Constructor function returns an instance of SedimentColumn with PorewaterProperty vectors of length n. Optionally provide values for Cl, O, and rho (otherwise values are undefined).

see also: PorewaterProperty, density

PorewaterDiffusion.LR04 — Method

LR04()

Generate a NamedTuple instance containing the Liesiecki & Raymo 2004 benthic stack (data, publication) interpolated for 1 ka timesteps and going forward in time from 5.32 Ma.

field	description

t	time (ka)

x	benthic δ¹⁸O (‰)

n	timesteps in t

PorewaterDiffusion.andrill1b — Method

andrill1b

Generate a CoreData instance with data from core ANDRILL-1B (from Pompilio+ 2007 (https://digitalcommons.unl.edu/andrillrespub/37)).

see also: CoreData

PorewaterDiffusion.andrill2a — Method

andrill2a

Generate a CoreData instance with data from core ANDRILL-2A (from Tracy+ 2010, doi:10.1130/G30849.1).

see also: CoreData

PorewaterDiffusion.boundaryconditions — Method

function boundaryconditions(Cl, d18O, x, sea2freeze,freeze2melt, meltrate, freezerate, Clsw, d18Osw, dz, dt)

Calculates sediment surface boundary condition for δ¹⁸O (d18O) and Cl⁻, based on the thermodynamic state described by the current current benthic δ¹⁸O value x and the threshold values corresponding to subglacial freezing sea2freeze and subglacial melting freeze2melt.

For melting or freezing states, calculates boundary condition from the assumed meltrate, freezerate, timestep dt, length-step dt, and composition of seawater Clsw and d18Osw.

PorewaterDiffusion.density — Method

density(chlorinity)

Calculates the density of a water parcel with chlorinity in units g/m³ (rather than kg/m³ for convenience with velocity)

see also: velocity

PorewaterDiffusion.diffuseadvectcolumn! — Method

diffuseadvectcolumn!(sc,k)

Calculate diffusive and advective transport of chlorinity and isotope-traced water through a sediment column described by properties in k, a NamedTuple generated by the Constants function.

Overwrites the o and p PorewaterProperty fields of sc – a SedimentColumn with pre-existing conditions of Cl, O, and rho in its o fields.

relies on: velocity, density, diffusionadvection,

see also: SedimentColumn, Constants

PorewaterDiffusion.diffusionadvection — Method

diffusionadvection(x,above,below,k1,k2,v,dt,dz)

Calculate the property of a node in a vertical profile given the combined effects of diffusion and advection. Returns the property given initial values for the node x, the overlying node above, the underlying node below, diffusion coefficients k1 and k2, vertical advection velocity v, timestep dt, and lengthstep dz. Alternatively provide the product of v * dt * dz for a minor speed-up.

PorewaterDiffusion.dt_climatetimestep — Method

dt_climatetimestep(katime,dt)

A helper function to calculate the number of diffusion model timesteps dt (in years) in each timestep of the climate timescale katime (in kiloannum). Returns an integer.

PorewaterDiffusion.getproposal — Method

getproposal(p, s)

Returns the value corresponding to field the field of Symbol s in Proposal instance p. Use in lieu of getproperty to avoid allocations.

PorewaterDiffusion.linterp — Method

linterp(x, x₁, Δx, y₂, y₁ )

Estimate the value of y corresponding to x given known coordinate (x₁, y₁), value y₂, and Δx= x₂ - x₁. Note that if x==x₁, y= y₁.

PorewaterDiffusion.loglikelihood — Method

loglikelihood( zₒ, μ, σ, zₘ, m )

Calculate the (relative) log-likelihood that model values in m simulated at depth nodes in zₘ were drawn from normally distributed observations with sample depths, mean values, and 1σ uncertainties in corresponding indices of the Vectors zₒ, μₒ, and σ. The simulated value is linearly interpolated between bounding depth nodes in zₘ.

see also: normll, linterp

PorewaterDiffusion.mcmurdoshelf — Method

mcmurdoshelf()

Generate a seawater NamedTuple with coretop porewater compositions from the McMurdo ice shelf. Pairs with core ANDRILL-1B.

see also: seawater, andrill1b

PorewaterDiffusion.mcmurdosound — Method

mcmurdosound()

Generate a seawater NamedTuple with southern McMurdo Sound seawater compositions. Pair with core ANDRILL-2A.

see also: seawater, andrill2a

PorewaterDiffusion.normll — Method

normll(x, μ, σ)

Calculate the (relative) log-likelihood that an observation x was drawn from the normal distribution with mean μ and standard deviation σ. If μ is a NaN (instead of missing for type homogeneity), returns 0.

Note: This function excludes a constant that will not vary for different vlaues of x to speed up calculation in metropolis. To calculate the absolute log-likelihood, take the log of normpdf

PorewaterDiffusion.normpdf — Method

normpdf(x, μ, σ)

Calculate the probability density of a normal distribution with mean μ and standard deviation σ at the value x.

PorewaterDiffusion.porewaterhistory! — Method

porewaterhistory!(sc::SedimentColumn, p::Proposal, k::Constants, climhist::NamedTuple, sw::NamedTuple, ka_dt::Int)

In-place version of porewaterhistory, which takes every input as an arg (rather than some defaults as kwargs). It also requires you to provide ka_dt – the number of diffusion timesteps in each thousand-year climate timestep.

see also: porewaterhistory

PorewaterDiffusion.porewaterhistory — Method

porewaterhistory(proposals [; k=Constants(), climatehistory=LR04(), seawater=mcmurdosound()])

Calculate the porewater advection-diffusion history of chlorinity and O-isotope-traced water in a sediment column described by properties in k (::Constants) over a given climatehistory (LR04 by default) and coretop seawater compositions.

The provided proposals (as custom type Proposal) describe the sensitivity and response of the system to climate fluctuations as recorded in climatehistory.

See diffuseadvectcolumn! for the underlying diffusion-advection transport calculations.

see also: porewaterhistory!, Proposal, Constants, LR04, seawater

PorewaterDiffusion.porewatermetropolis — Method

porewatermetropolis...

Not tested, yet...

PorewaterDiffusion.stopwatch — Method

stopwatch(i, n, clock)

Convenience function for [porewatermetropolis] that returns a String reporting the progress at step i for total steps n with start time t (in s since the epoch).

PorewaterDiffusion.strictpriors — Method

strictpriors(p, record_max_age, climatelimits)

Evalute strict constraints on priors that will automatically reject a proposal:

Onset date beyond the climate record timespan (in ka)

Negative freezing or melting rates.

Nonphysical subglacial thresholds – melting at lower benthic δ¹⁸O than freezing or values exceeding the record extrema.

PorewaterDiffusion.velocity — Method

velocity(x, above, k)

Calculate the velocity (m/yr) at a node with density x, given the density of the node above, and the hydraulic conductivity k (m/yr).

see also: density